+++ src/icos_config/index.js
@@ -2,15 +2,19 @@ import moment from 'moment';
 
 const requireAll = r => r.keys().map(fileName => require(`./${fileName.replace('./', '')}`).default);
 
-const icosAsList = requireAll(require.context('./', true, /\.js$/));
+export const getICOsAsList = () => {
+  const icosAsList = requireAll(require.context('./', true, /\.js$/));
+  icosAsList.sort((a, b) => moment(b.addingDate, 'DD-MM-YYYY').toDate() - moment(a.addingDate, 'DD-MM-YYYY').toDate());
+  return icosAsList;
+};
 
-icosAsList.sort((a, b) => moment(b.addingDate, 'DD-MM-YYYY').toDate() - moment(a.addingDate, 'DD-MM-YYYY').toDate());
-
-const icosAsDict = {};
-icosAsList.forEach((element) => {
-  if (element) {
-    icosAsDict[element.crawdSaleTokenContract] = element;
-  }
-});
-
-export default icosAsDict;
+export const getICOsAsDict = () => {
+  const icosAsDict = {};
+  const icosAsList = getICOsAsList();
+  icosAsList.forEach((element) => {
+    if (element) {
+      icosAsDict[element.crawdSaleTokenContract] = element;
+    }
+  });
+  return icosAsDict;
+};
